Feat/admin parties promises

+++ lib/database/parties/deleteParty.ts
@@ -3,13 +3,8 @@ import { supabaseAdmin } from "@/lib/supabaseClient";
 
 const log = new Logger();
 
-export const deleteParty = async (name: string) => {
-  const { data: result, error: insertError } = await supabaseAdmin
-    .from("parties")
-    .delete()
-    .eq("name", name)
-    .select("id")
-    .single();
+export const deleteParty = async (id: number) => {
+  const { data: result, error: insertError } = await supabaseAdmin.from("parties").delete().eq("id", id).select("id").single();
 
   if (insertError) {
     log.error(`Error deleting party:`, insertError);
