+++ lib/database/parties/subjects/deleteSubject.ts
@@ -3,13 +3,8 @@ import { supabaseAdmin } from "@/lib/supabaseClient";
 
 const log = new Logger();
 
-export const deleteSubject = async (name: string) => {
-  const { data: result, error: insertError } = await supabaseAdmin
-    .from("subjects")
-    .delete()
-    .eq("name", name)
-    .select("id")
-    .single();
+export const deleteSubject = async (id: number) => {
+  const { data: result, error: insertError } = await supabaseAdmin.from("subjects").delete().eq("id", id).select("id").single();
 
   if (insertError) {
     log.error(`Error deleting subject: `, insertError);
